refactor: export attribute types from higher level
