+++ packages/consumption/src/modules/attributes/index.ts
@@ -1,6 +1,13 @@
 export * from "./AttributesController";
 export * from "./events";
 export * from "./local/AttributeTagCollection";
-export * from "./local/attributeTypes";
+export * from "./local/attributeTypes/AttributeWithPeerSharingDetails";
+export * from "./local/attributeTypes/ForwardableAttribute";
+export * from "./local/attributeTypes/LocalAttribute";
+export * from "./local/attributeTypes/OwnIdentityAttribute";
+export * from "./local/attributeTypes/OwnRelationshipAttribute";
+export * from "./local/attributeTypes/PeerIdentityAttribute";
+export * from "./local/attributeTypes/PeerRelationshipAttribute";
+export * from "./local/attributeTypes/ThirdPartyRelationshipAttribute";
 export * from "./local/sharingDetails";
 export * from "./local/successorParams";
