+++ packages/consumption/src/modules/requests/itemProcessors/readAttribute/ReadAttributeRequestItemProcessor.ts
@@ -20,6 +20,7 @@ import { nameof } from "ts-simple-nameof";
 import { ConsumptionCoreErrors } from "../../../../consumption/ConsumptionCoreErrors";
 import {
     AttributeSucceededEvent,
+    LocalAttribute,
     OwnIdentityAttribute,
     OwnIdentityAttributeSuccessorParams,
     OwnRelationshipAttribute,
@@ -30,7 +31,6 @@ import {
     ThirdPartyRelationshipAttribute,
     ThirdPartyRelationshipAttributeSuccessorParams
 } from "../../../attributes";
-import { LocalAttribute } from "../../../attributes/local/attributeTypes";
 import { ValidationResult } from "../../../common/ValidationResult";
 import { GenericRequestItemProcessor } from "../GenericRequestItemProcessor";
 import { LocalRequestInfo } from "../IRequestItemProcessor";
